requirements.txt
streamlit>=1.35.0
pandas>=2.2.2
openpyxl>=3.1.2
xlrd>=2.0.1
python-docx>=1.1.0
PyPDF2>=3.0.1
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
